Construction_data_lake_data_warehouse/application/requirements.txt
kafka-python-ng
mysql-connector-python
apache-beam
schedule